Add the concept of Nostalgia versions
Nostalgia versions are versions of OrionCraft that were created explicitly for nostalgia reasons.

A good example of this is the 1.5.2 and 1.6.4 versions, which exist not because people use them, but because of nostalgia reasons.
No one will obviously use them, but what matters is what we feel when working on them, the memories we can bring back.